refactor: Extract utils
